Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
pydantic[email]
httpx
python-jose
rapidfuzz
slowapi
pytest